refactor with new lint rule
