+++ apps/dashboard/analysis_options.yaml
@@ -1,4 +1,4 @@
-include: package:flutter_lints/flutter.yaml
+include: package:pedantic_mono/analysis_options.yaml
 linter:
   rules:
 analyzer:
