+++ apps/dashboard/lib/src/features/workflow/presentation/workflow_page.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:dashboard/src/common_widgets/margins.dart';
 import 'package:dashboard/src/extensions/build_context_extension.dart';
 import 'package:dashboard/src/features/workflow/domain/workflow_model.dart';
+import 'package:dashboard/src/features/workflow/presentation/edit_workflow_dialog.dart';
 import 'package:dashboard/src/features/workflow/presentation/workflow_edit_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
@@ -19,7 +20,7 @@ class WorkflowPage extends StatelessWidget {
         centerTitle: true,
         actions: [
           IconButton(
-            onPressed: () async => await FirebaseAuth.instance.signOut(),
+            onPressed: () async => FirebaseAuth.instance.signOut(),
             icon: const Icon(Icons.logout),
           ),
           horizontalMargin20,
@@ -35,8 +36,10 @@ class WorkflowPage extends StatelessWidget {
             return const Center(child: CircularProgressIndicator());
           }
           final workflows = snapshot.data!.docs
-              .map((e) =>
-                  WorkflowModel.fromJson(e.data() as Map<String, dynamic>))
+              .map(
+                (e) =>
+                    WorkflowModel.fromJson(e.data()! as Map<String, dynamic>),
+              )
               .toList();
           return SingleChildScrollView(
             child: Column(
@@ -48,8 +51,8 @@ class WorkflowPage extends StatelessWidget {
                   itemBuilder: (_, index) {
                     final workflow = workflows[index];
                     return Padding(
-                      padding: const EdgeInsets.all(16.0),
-                      child: WorkflowCard(workflow: workflow),
+                      padding: const EdgeInsets.all(16),
+                      child: WorkflowCard(workflow: workflow, index: index),
                     );
                   },
                 ),
@@ -58,7 +61,6 @@ class WorkflowPage extends StatelessWidget {
                   style: ElevatedButton.styleFrom(
                     side: BorderSide(
                       color: context.primaryColor,
-                      width: 1,
                     ),
                   ),
                   onPressed: () async =>
@@ -77,9 +79,9 @@ class WorkflowPage extends StatelessWidget {
     BuildContext context, {
     required WorkflowModel workflow,
   }) {
-    showDialog(
+    showDialog<void>(
       context: context,
-      builder: (context) => WorkflowDialog(workflow),
+      builder: (context) => EditWorkflowDialog(workflow),
     );
   }
 }
