+++ apps/dashboard/pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   json_annotation: ^4.9.0
   riverpod_annotation: ^2.5.3
   signals_flutter: ^5.5.0
+  uuid: ^4.5.1
 dev_dependencies:
   build_runner: ^2.4.13
   custom_lint: ^0.6.8
@@ -27,6 +28,7 @@ dev_dependencies:
     sdk: flutter
   freezed: ^2.5.7
   json_serializable: ^6.8.0
+  pedantic_mono: ^1.28.0
   riverpod_generator: ^2.4.3
   riverpod_lint: ^2.3.13
 flutter:
